+++ ChangeLog.md
@@ -9,6 +9,7 @@
     - Define a new 'onRebuildMenuItems()' hook called each time connection status or unverified email address count change
         Replace and obsoletes namedDropdownItems()
     - Obsolete dropdownItems() and dropdownItemsExt() unused functions
+    - menuItems() private function refactoring to better handle the reactivity
 
 ### 2.2.0